requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
SQLAlchemy==2.1.0
pydantic==1.10.11
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.6
